Add ability to download route information from an API server
This collects the callsigns in question in batches - the first one
after 5 seconds, after that every 10 seconds, and calls a (configurable)
backend server. This server can be run locally or a public server can be
used - right now this is configured for api.adsb.lol.

The feature is turned off by default.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>